Stop using names instead of ids to distinguish chapters and section
This is part of removing named_path since it didn't make any sense and
was redundant with the full path.

Also properly handle the case where a video doesn't have a section or
chapter.

While I was here I also did some things that made the PR longer but
simpler:

In particular, I:
- Factored OEXHelperVideoDownload since most of it was
redundant with OEXVideoSummaryList (which I renamed to OEXVideoSummary
since it wasn't a list).
- Introduced a type for path entries and started using that.
- Removed several properties (ChapterName and SectionName) that
  encouraged using names where we shouldn't.
- Fixed a couple lines that prevented us from enabling warnings as
  errors.

JIRA: MA-258