Prioritize google profile email address
There's a case where the google profile info isn't coming through. I
can't reproduce it, but it seemed like we should at least be able to get
an email address as part of the google authenitcation, so this tries to
use that email first. It's more likely to be the primary email address
anyway.

JIRA: https://openedx.atlassian.net/browse/MA-390