UI for offline mode.
May need to add to other places appropriate. Visuals are not final
(per Marco, trying to do something that takes up less space than the one in the
spec, but haven't reviewed it yet).

Adds OfflineModeController to make it easy for any controller to implement a
standard offline mode view.

Adds ContentInsetsController to make it easy to update a controller's
edge insets as things like offline mode affect it.

https://openedx.atlassian.net/browse/MA-746