Switch to modern status bar APIs.
As of iOS 7, view controllers should be responsible for their status bar
style/visibility instead of it being a global property. This adopts
those new APIs.

This was prompted by the FullScreenMessageViewController having a
visible status bar that happened to be hidden because it had a white
background. Fixes that issue as well as the text not wrapping on that
screen.

This also adopts the standard navigation bar for
OEXFrontCourseViewController so that the push/pop animation looks
normal.

JIRA: https://openedx.atlassian.net/browse/MA-939?filter=13805