Update to swift 2
This includes migrating all of our code and also third party libraries.
We should do a better job of getting on the latest version of those
third party libraries once we drop iOS 7 support and we can pick up
frameworks. Alamofire got refactored so that might end up being some
work. But we don't use most of it, so I'm hoping it's not much of a
problem.

I also had to weak some xib usage to account for changes on iOS 9. In
particular, it appears that the system will now automatically try to
load xibs named "SomeView" for the class named "SomeViewController".
This was wreaking havoc on some of the dicussions code where for no
particularly good reason, it had a XIB which was the full content of the
view, but loaded using loadNibNamed, rather just through the view
controller nib loading mechanism.

I tried to clean up all the warnings too, but there were a few coming in
from other sources I couldn't get rid of. Unfortunately, the swift
compiler doesn't seem to have a warnings as errors flag, but we should
not add any more warnings.

Further, circleci seems to just be borked, so I'm switching us back to
travis.