Fix issue with top message headers always appearing at bottom of screen
This appears to have been caused by some sort of weird change to how
view controller layout guides are calculated on iOS 9.

It turned out the easiest way to fix this was to start using the new
layout anchor's API, which easiest way to fix was to upgrade SnapKit,
which easiest way to do was to drop iOS 7 support.

This change drops iOS 7 support+updates SnapKit+fixes switches to layout
anchors where available.