Fix issues with setting up listeners in video block and handouts.
In handouts we were only setting the listener to listenOnce which means
it will only send the cached version.

In the video block code we were adding a listener each time we reloaded
the stream. In practice we were only loading the stream once so it
didn't matter, but if we wanted to add a refresh feature this could lead
to a bug.