Update assessments screen analytics events.
Previously, I added the events to match what we had previously, but they
were missing some screens and important data. This makes the following
changes:

Extends screen events so that they take an optional courseID and an
extra value (that will display on GA's 'action' field.
Fires an additional analytics event with extra of information of
category "screen" when firing a screen event so we can get extra
information to GA. Their screen events just have a name as the only
property you can send.

Updates the following screens:

```
    Course List:
        Label: "Course Dashboard"
        Action: none
    Course Outline:
        Label: "Course Outline"
        Action: none
    Section Outline:
        Label: "Section Outline"
        Action: <section-name>
    Scrolling Unit Detail:
        Label: "Unit Detail"
        Action: <unit-name>