Fix issue with bunched text.
We were rotating a view 90 degrees to get its icon to be rotated 90
degress which screwed up autolayout. This switches to actually rendering
the icon rotated. Since this was a thing we were already doing in a one
off way, this generalizes the concept and creates a rotated icon
renderer. It also factors the icon renderer list to better catch
exhaustivity when not using FontAwesome.

JIRA: https://openedx.atlassian.net/browse/MA-1440