Show video speed/caption controls in landscape.
These were getting covered up because we were adding the controller
directly to the front window. This change has us move to the front of
the root view controller. This is still not a great fix since the alerts
come in sideways, but it's an improvement over not showing them at all.
Unfortunately, the correct fix for this is either to:
1. Restore our more custom controls and add them as a subview of the
video player.
2. Significantly refactor the video player and its assorted containers
to handle landscape in a more disciplined way (migrate to
AVPlayerViewController and have the container be responsible for
fullscreen support + showing/hiding status bar, toolbar, and navigation
bar.

Also includes minor cleanup:
- Use system notification symbol instead of quoted string
- Remove unused method

JIRA: https://openedx.atlassian.net/browse/MA-1446