Update design of course card.
Per product:
- Decrease font size slightly.
- Remove weird footer on course list.

While here I switched the dashboard to use a stack view, so that we can
naturally pick up card geometry changes via autolayout and make it
easier to add a spacer that's not part of the card itself.

This necessitated improving the autolayout behavior of the course card
itself.

Also renamed the course card to reflect that it's a course card and not
just used in the dashboard.