Add basic auth credentials to config.
- Connect basic auth credentials to NetworkManager.

Naively it might seem that you can just add credentials to the global
NSURLCredentialStorage. Unfortunately, basic auth has a notion of a
realm and those need to match in order for the protection spaces to
match. So we'd either have to add that to the config or do what we're
doing. Further, we have to go this route anyway, since WKWebView doesn't
pick up the global credential storage.

JIRA: https://openedx.atlassian.net/browse/MA-1570