Prevent long email addresses + names from getting hidden
There's a bunch of space on the right of this for the overlap with the
overlaying content. This shrinks the name and email to fit within that
area.

It also tweaks the bottom constraint for the email address so it
provides more even vertical spacing on extra long email addresses.