Remove new navigation flag.
This flag is now necessary to run and we're about to release it. So
remove the old thing.

While trying to fix this, I noticed a bug with video controls that this
removal would cause to be weirder. We only don't support next/prev from
the video block (since it doesn't work right in landscape), but we were
hiding it everywhere (and then confusingly sometimes showing it if the
right notification fired). This provides a simple on off switch that
should always work.