Begin cleanup of environments
The goal of this is to make dealing with environments significantly less
boilerplatey. There are two angles on this:

1. Make it easier to make up an empty test environment with a few
overrides for common cases.

This works by having a standard RouterEnvironment with default
implementations that less us swap in specific versions of only the
dependencies we care about.

2. Support using protocol composition to describe a class's environment
and then just pass around the router environment suitably upcast for the
current context.

For example:

```
class SomeController {
    typealias Environment = protocol<NetworkManagerOwner, OEXSessionOwner>
}

// In OEXRouter
let controller = SomeController(environment : self.environment)
```

instead of

```
class SomeController : UIViewController {
    struct Environment {
        let networkManager: NetworkManager
        let session : OEXSession

        init(networkManager : NetworkManager, session: OEXSession) {
            self.networkManager = networkManager
            self.session = session
        }
    }
}

// In OEXRouter:
let environment = SomeController.Environment(...)
let controller = SomeController(environment: environment)
```

This does not migrate existing controllers or add owning protocols for
all of our standard dependencies. I expect that to be a more gradual
migration as appropriate.