Fix unit tests
We had a listener which was never getting removed. It should really just
be automatically getting removed once it gets dealloced. So this probably
indicates a memory cycle in the tests somewhere. I took a pass at the
actual app and that leak doesn't seem to be there, so I don't think it's
a serious problem - though we should fix it.

But in the mean time this fixes the tests.