Add KVO helpers.
KVO is a terrible API for a number of reasons. This wraps so its block
based and you don't have to worry about dispatch or removing listeners.

I don't really like KVO, but sometimes it's the only way to listen to
framework properties.

Particular use case (UIScrollView contentSize) comes next.