Introduce pagination primitives
The goal here is to get rid of NetworkPaginator and replace it with
PaginationController which is more general, testable, and has a simpler
interface

Next steps are:
1. Adopt PaginationController in CourseCatalogViewController.
2. Replace NetworkPaginator where it is currently used in the
discussions code.

At that point we'll have one system where the differences in payload
content (whether there's a "pagination" package or not) are abstracted
into the specific Paginator implementation used.