Add non persistence support to remote image.
We don't want the course catalog to persist all the course images.

Factor CourseCardViewModel to better break apart the different cases.
This also makes the environment part of the model setup instead of the view,
which seems like a more natural separation of concerns and should make
it harder to forget, which we've had a couple bugs about.