Start deprecation of OEXStatusMessageViewController.
Provide standard way to show an overlay message on a view controller.
Does *not* update all messages to be in that style or use that API. In
particular, some of the Objective-C based screens still use the weird
custom nav bar stuff and so depend on the extra stuff
OEXStatusMessageViewController is doing. And the error in the profile
editor was sufficiently different in style and behavior that I didn't
want to mess with it at at this moment.

This new implementation doesn't rely on a singleton, supports custom
views displayed in a standard way, and is desgined around autolayout.

This is really just going to be used to make it easier to show the
large offline info message when we modernize the course list.