Introduce enrollment manager
This introduces a new class ``EnrollmentManager`` which is responsible
for keeping track of the user's current enrollments. It will replace
uses of OEXInterface for the same purpose as part of removing uses of
that class.

Currently the app depends on the main course list screen to actually load the
enrollments. This way we can load enrollments without involving any UI
classes.