Fix scrolling for web view in course detail
The way this used to work was there was a scrollview that contained a
bunch of content views and a webview at the bottom. The webview was
resized to fit its content and had scrolling disabled.

Unfortunately, I tried this with non-trivial content (i.e. not just my
test content), it had tons of layout issues. This fixes it by removing
the container scrollview, having a constant height webview, and putting
the non web content at the top of it, offsetting it as the webview
scrolls and using content insets to offset the web content below the non
web content.

JIRA: https://openedx.atlassian.net/browse/MA-1851