Fix choice of rendered_body over raw_body.
So when I went to go refactor the forum models, I made a mistake and
picked "rendered_body" instead of "raw_body", since I figured "raw_body"
would contain markdown/html/whatever. This is true. However, rendered_body
*always* contains HTML, in particular, every string is wrapped in <p>
tags.

On closer investigation, it turns out that actually neither of these is
right and we don't have a processed plain text version of the post
content. This is a problem and I'm looking into it, but in the mean
time we should use raw_body since it *usually* doesn't have stuff in it.

JIRA: https://openedx.atlassian.net/browse/MA-1879