Fix crash with KVO lifecycle
KVO has a stupid system where you have to remove an observer before the
observed object gets deallocated. We were being too clever about weak
references and so the observed object could get deallocated before the
observer. Ideally, we'd just nuke the listener in that case, but we
can't trigger an event *just before* the observed object gets
deallocated (at least without swizzling, which seems excessive).

So instead, we update the pattern to explicitly retain the observer.
This is typically how lifetimes work anyway, so it shouldn't be an
issue.

Unfortunately, I couldn't actually reproduce the related crash, but it's
clear from the stack trace that this is the underlying issue. The test I
added crashes on the old implementation.

JIRA: https://openedx.atlassian.net/browse/MA-2136