Introduce edXCore lib
We've been talking about breaking our code up into libraries for a
while. This is the first step. It was necessary to do since XCUITest
runs in a separate process and so if we wanted to share any code between
the UI tests and the app we needed a common framework. Since we wanted
to do this anyway, this seems like a good place to start.

This also establishes a structure for future sub-libraries:

Source/<LibraryName>/SupportingFiles
Source/<LibraryName>/Code
Source/<LibraryName>/Tests

The purpose of Core is for things that would make sense in *any*
application and aren't dependent on UIKit.

Finally, I moved a couple things into Core that I knew I would need for
the UI tests.
Particularly,
JSON, NSBundle+OEXConveniences and NSString+OEXFormatting

I expect to soon add an edXAPI library that is basically just the API
methods and the associated models.