Add a login end to end test based off the config
This is one of those commits that grew in the telling and I'm sort of
embarrassed about it. The main thing that needed to happen here was to
get NetworkManager and OEXConfig into Core.

This was so we could make an API call from the end to end tests, which
don't link directly to the edX code. This in turn involved moving a
bunch of transitive dependencies and minor factoring thereof.

In particular, we had to remove dependencies on OEXFileUtility and
OEXRouter, since those only exist at the app level. We also pulled
apart several categories and things that had parts that only made
sense at the app level.

Further, the registration endpoint doesn't support JSON bodies, so we
had to add support for form encoding.