Prefer fallback video to youtube
When we added the youtube redirect profile, we prioritized that ahead of
the fallback video (unprocessed video of unknown size). This seemed kind
of reasonable since some of those videos are quite large and don't work
well on mobile. But we're getting user complaints about offline videos
vanishing and obviously the youtube videos obviously don't work offline,
so swap this back around. This is also more consistent with what Android
is doing.

JIRA: https://openedx.atlassian.net/browse/MA-2256