Fix crash reading cache entries.
This is pretty unfortunate. When we moved the PersistentResponseCache
from the "edX" module to the edXCore module, it broke unarchiving cache
entries - since NSCoding looks things up by class name and being in a
different module secretly gives you a different class name - even though
that's not visible to the user.

This change adds some rerouting logic for the unarchiver to better
handle missing classes and route the cache entries to a class that
actually exists.