Better control auth usage in RemoteImage
We were encountering an issue where any image request using RemoteImage would send a
user's oauth credentials as part of the request. This is bad for two
reasons:

1. It's a security issue.
2. Some sources - like AWS - will actually reject clients with bad
tokens even if the underlying asset does not require authorization

I don't believe this is a short term security emergency because we only
request URLs via remote image that have been laundered through the LMS.
But we should ensure this is fixed before adding any more uses.

This change limits sending credentials to only requests to the host under API_BASE_URL.