Using GoogleService-Info.plist to Configure Firebase SDK  (#1269)
* Remove Firebase custom initialization.
An option is available to Initialize the Firebase with `configureWithOptions` but if it is initialized with `configureWithOptions` then some of the Firebase KITS didn't work as expected. Especially Analytics didn't work well.
More details about Analytics issue  can be found here https://github.com/firebase/quickstart-ios/issues/75

* Update build.gradle script to add Firebase keys to the GoogleService-Info.plist
We can't live without adding GoogleService-Info.plist files to the project because it is required. But as it contains sensitive information and we can't risk exposing sensitive information.
I have removed all the sensitive information from the GoogleService-Info.plist file and updated the build.gradle in a way that it will read Firebase keys from the config file and will add them to the GoogleService-Info.plist at the build time.