Course Video Indexing on the CourseID
In OEXInterface courseVideos indexing was based on course_outline but a TODO was there to update the courseVideos indexing on course_id.
If a course video was downloading and user kills the app then on restarting the app, downloading on the underlying video restarted but downloading struck on 1-5% and never completed.
This PR address the video restart issue as well.