Add support of Xcode 10.3 and Swift 5 (#1298)
* Add support of Xcode 10.3 and Swift 5 by running code convertor tool of Xcode
Updating Snapkit for Swift 5
Removing warning which were appearing because of Swift 5 upgrade

* Removing warning and Updating current and previous OS version for Tests

* Update reference images for screenshot tests.
We were using reference images of iOS 9 & iOS 10  for snapshot tests but Travis-ci is no longer providing support of iOS 9 with Xcode 10.3. The least supported version of iOS with Xcode 10.3 configuration is 10.3.1.
I have updated the references images from IOS 9 & iOS 10 to iOS 11.4 & iOS 12.4. I have selected the highest versions so that we don't have to update them in the near future.