Always play AVPlayer by setting the rate directly.
Remove the previous method of observing when the player changed state and setting the rate afterward.
There is an issue in `resume(at:)` that resulted in `player.play()` being called, but a play status change either not happening or not being observed.  This meant that the rate was sometimes not set after play began with the default 1.0 rate.