refactor: Update how version numbers are used
"Build Version" aka "Bundle Version" or "Current Project Version" is intended to be an internal version.
"Short Version" aka "Marketing Version" is intended to be what is typically displayed to and understood by users.

We were using "Build Version" as the user-understood version, and "Short Version" as just the Major.Minor part of the "Build Version".  This change replaces our use of "Build Version" with "Short Version" in most places, and adds a new "Display Version" that is a concatenation of "Build Version" and "Short Version" for use when displaying the version to users in the app in Login, What's New, and Account screens, as well as for User Agent strings.

IMPORTANT: This change is intended to happen alongside updated usage of "Version" and "Build" strings in the edX target settings.  That is, moving forward, the "Version" string - as the "Short Version" - should be the version that we intend users to see in the App Store, and general understand to be the app version.  The "Build" string should be used to identify exactly which commit/build the app is.