build: Remove CoreBluetooth.framework as a dependency
CoreBluetooth.framework is not used in the app, so we do not need to link to this framework.  Simply linking to it causes Apple to warn that we do not have a NSBluetoothAlwaysUsageDescription in our info.plist when we upload new binaries to App Store Connect, so this will also prevent that warning.

* Remove CoreBluetooth.framework from linked binaries in the project file.
* Delete CoreBluetooth.framework from the Frameworks folder